♻️ refactor: move DeepModelingToggle to SessionForm and add MicButton to GitHubSessionFormPresenter
- Move DeepModelingToggle from UI package to SessionForm components
- Update imports to use local component instead of UI package
- Add MicButton next to DeepModelingToggle in GitHubSessionFormPresenter
- Update both components to use React 19 ref pattern without forwardRef
