+++ frontend/packages/ui/src/components/index.ts
@@ -22,5 +22,4 @@ export * from './Tooltip'
 export * from './Modal'
 export * from './Input'
 export * from './ContextMenu'
-export * from './DeepModelingToggle'
 export * from './Spinner'
